+++ lib/shared/constants.dart
@@ -1,0 +1,3 @@
+class Constants {
+  static const String starWarsApiUrl = 'https://swapi.dev/api/people/';
+}
